code cleanup: moved database exceptions and room classes to separate packages
