+++ src/io/github/sspanak/tt9/db/exceptions/DictionaryImportAlreadyRunningException.java
@@ -1,4 +1,4 @@
-package io.github.sspanak.tt9.db;
+package io.github.sspanak.tt9.db.exceptions;
 
 public class DictionaryImportAlreadyRunningException extends Exception{
 	public DictionaryImportAlreadyRunningException() {
