+++ src/io/github/sspanak/tt9/db/room/T9RoomDb.java
@@ -1,9 +1,9 @@
-package io.github.sspanak.tt9.db;
+package io.github.sspanak.tt9.db.room;
 
 import androidx.room.Database;
 import androidx.room.RoomDatabase;
 
 @Database(version = 5, entities = Word.class, exportSchema = false)
-abstract class T9RoomDb extends RoomDatabase {
+public abstract class T9RoomDb extends RoomDatabase {
 	public abstract WordsDao wordsDao();
 }
