+++ src/io/github/sspanak/tt9/db/room/WordsDao.java
@@ -1,4 +1,4 @@
-package io.github.sspanak.tt9.db;
+package io.github.sspanak.tt9.db.room;
 
 import androidx.room.Dao;
 import androidx.room.Insert;
@@ -9,12 +9,12 @@ import java.util.ArrayList;
 import java.util.List;
 
 @Dao
-interface WordsDao {
+public interface WordsDao {
 	@Query("SELECT COUNT(id) FROM words WHERE :langId < 0 OR lang = :langId")
 	int count(int langId);
 
 	@Query("DELETE FROM words WHERE LANG IN(:langIds)")
-	int deleteByLanguage(ArrayList<Integer> langIds);
+	void deleteByLanguage(ArrayList<Integer> langIds);
 
 	@Query("SELECT COUNT(id) FROM words WHERE lang = :langId AND word = :word")
 	int doesWordExist(int langId, String word);
